backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
asyncpg==0.29.0
aiosqlite==0.19.0
psycopg2-binary==2.9.9
alembic==1.13.1
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
httpx==0.25.2
email-validator==2.1.0
passlib[bcrypt]==1.7.4
argon2-cffi==23.1.0
PyJWT==2.8.0
transformers==4.36.0
torch==2.1.0
sentencepiece==0.1.99
